fix uname detection logic on windows
This was breaking mkl-devel specifically, generating incorrect site.cfg files.
We made a one-off split of this recipe just for mkl-devel, to avoid need to
update mkl package also.  This is for historical reasons - previous mkl
packages were built with conda 3.0.x, and their hashes would change under
conda-build 3.1.x+.  This commit is here to be used for the next MKL update,
2018.0.2.